解bug -- 颜色闪烁问题(错把progress定义为NSInteger了, 应该为CGFloat, 我是个粗心的马大哈 😓)

+++ MusicPlayer/Base.lproj/View/HBLyricView.h
@@ -29,6 +29,6 @@
 /// 当前播放歌词索引
 @property (nonatomic, assign) NSInteger currentLyricIdx;
 /// 当前播放歌词进度 (设置颜色)
-@property (nonatomic, assign) NSInteger progress;
+@property (nonatomic, assign) CGFloat progress;
 
 @end
